datapath: Move node_addressing.go to pkg/datapath
The code is no longer specific to Linux as it is implemented
in terms of queries against the LocalNodeStore and the devices
table.
