+++ pkg/datapath/node_addressing.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package linux
+package datapath
 
 import (
 	"context"
